henter saksnummer og sender det med når søknad sendes til saksbehandling-api
