+++ src/main/kotlin/no/nav/tiltakspenger/soknad/api/soknad/NySøknadCommand.kt
@@ -29,6 +29,7 @@ data class NySøknadCommand(
             journalpostId = null,
             opprettet = innsendingTidspunkt,
             eier = eier,
+            saksnummer = null,
         )
     }
 }
