+++ src/main/resources/db/migration/V3__soknad_saksnummer_kolonne.sql
@@ -1,0 +1,2 @@
+alter table søknad
+  add column saksnummer varchar;
